Fix ruff line-length errors and mypy type errors on this branch
Wraps 7 lines over ruff's 100-char limit across
backend/{holdings,transactions}/tests.py, backend/market_data/views.py,
backend/transactions/views.py, and packages/core's transactions.py/
user_profiles.py.

Also fixes 19 pre-existing mypy errors in packages/core, all in
tests/test_client.py plus one narrow fix in client.py itself:
- get_news/get_profile/get_events/get_dividends return
  dict[str, Any] | None - every test indexing the result straight away
  now asserts it's not None first.
- TestGetCatalog's `expected` dict is now annotated dict[str, Any] so
  .update() with a mixed-type row doesn't narrow it to dict[str, None].
- MarketDataClient.__init__'s cache_ttl_seconds widened from int to
  float - a sub-second TTL (0.05s) is what the cache-expiry test
  actually needs to stay fast, and the value is only ever used in
  arithmetic/comparisons, never anywhere that requires an int.

Verified: ruff and mypy both clean across backend/ and packages/core/;
100 packages/core tests and 113 backend tests (holdings, market_data,
transactions) still pass.

+++ packages/core/src/equicast_core/client.py
@@ -190,7 +190,7 @@ class MarketDataClient:
         bucket: str,
         s3_client: Any = None,
         region_name: str | None = None,
-        cache_ttl_seconds: int = DEFAULT_CACHE_TTL_SECONDS,
+        cache_ttl_seconds: float = DEFAULT_CACHE_TTL_SECONDS,
     ) -> None:
         self._bucket = bucket
         self._s3 = s3_client or boto3.client("s3", region_name=region_name)
